requirements.txt
caldav
envelope
flask
icalendar
python-dateutil
requests
mininterface[basic]